refactor(holidays): Using model types
