+++ src/app/components/holidays/OmerPrompt.tsx
@@ -2,10 +2,10 @@ import React from 'react';
 import { TouchableWithoutFeedback, View } from 'react-native';
 import { connect } from 'react-redux';
 
+import { IHoliday } from '../../../core/models/holidays';
 import { getHolidays, getNow } from '../../../core/store/get';
 import { OmerPromptText } from '../../Styles';
 import getOmer from '../omer/utilities/getOmer';
-import { IHoliday } from './types';
 
 interface IExternalOmerPromptProps {
 	goToOmerPage: () => void;
