+++ src/app/components/holidays/utilities/sorters.ts
@@ -1,4 +1,4 @@
-import { IHoliday } from '../types';
+import { IHoliday } from '../../../../core/models/holidays';
 
 export const ascending = (h1: IHoliday, h2: IHoliday): number =>
 	h1.date.diff(h2.date).shiftTo('milliseconds').milliseconds;
